requirements.txt
tensorflow==2.5.0
pillow
streamlit
numpy
